backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic-settings==2.7.1
motor==3.6.0
python-dotenv==1.0.1
dnspython==2.7.0
google-auth==2.36.0
PyJWT==2.10.1
